added functional tests for login and register page
